+++ arcade.city/components/no
@@ -1,0 +1,101 @@
+/*
+auto-generated by: https://github.com/react-spring/gltfjsx
+author: 3DLadnik (https://sketchfab.com/3DLadnik)
+license: CC-BY-4.0 (http://creativecommons.org/licenses/by/4.0/)
+source: https://sketchfab.com/3d-models/3d-printable-the-three-graces-58e0ae19e2984b86883edc41bf43415a
+title: 3D Printable The Three Graces
+*/
+
+import React, { useRef } from 'react'
+import { useFrame } from '@react-three/fiber'
+import { useGLTF } from '@react-three/drei'
+import lerp from 'lerp'
+
+export default function Model(props) {
+  const group: any = useRef()
+  const dis = useGLTF('/graces-draco.glb')
+  console.log(dis)
+  const wat = false
+  const nodes = false
+  console.log(wat)
+
+  useFrame(({ clock, mouse }) => {
+    if (!nodes || !group || !group.current) return
+    group.current.rotation.y = lerp(
+      group.current.rotation.y,
+      mouse.x * (Math.PI / 5),
+      0.005
+    )
+  })
+
+  if (!wat) return <></>
+
+  return (
+    <group {...props} dispose={null}>
+      <group ref={group}>
+        <mesh
+          castShadow
+          receiveShadow
+          geometry={nodes.Node_3.geometry}
+          rotation={[-Math.PI / 2, 0, 0]}
+          scale={[0.2, 0.224, 0.224]}
+        >
+          <meshStandardMaterial
+            attach='material'
+            roughness={0.9}
+            metalness={0.5}
+            color='#474747'
+          />
+        </mesh>
+        <Lights />
+      </group>
+    </group>
+  )
+}
+
+function Lights() {
+  const groupL = useRef()
+  const groupR = useRef()
+  const front = useRef()
+
+  useFrame(({ clock, mouse }) => {
+    groupL.current.rotation.y = lerp(
+      groupL.current.rotation.y,
+      -mouse.x * (Math.PI / 2),
+      0.1
+    )
+    groupR.current.rotation.y = lerp(
+      groupR.current.rotation.y,
+      mouse.x * (Math.PI / 2),
+      0.1
+    )
+    front.current.position.x = lerp(front.current.position.x, mouse.x * 12, 0.4)
+    front.current.position.y = lerp(
+      front.current.position.y,
+      7 + mouse.y * 4,
+      0.4
+    )
+  })
+
+  return (
+    <>
+      <group ref={groupL}>
+        <pointLight position={[0, 7, -15]} distance={15} intensity={10} />
+      </group>
+      <group ref={groupR}>
+        <pointLight position={[0, 7, -15]} distance={15} intensity={10} />
+      </group>
+      <spotLight
+        castShadow
+        ref={front}
+        penumbra={1}
+        angle={Math.PI / 3}
+        position={[0, 0, 8]}
+        distance={11}
+        intensity={8}
+        shadow-mapSize-width={2048}
+        shadow-mapSize-height={2048}
+      />
+    </>
+  )
+}
