+++ arcade.city/next-env.d.ts
