+++ arcade.city/tsconfig.json
