+++ package.json
@@ -16,8 +16,7 @@
     "dev:arcade.city": "yarn workspace @arcadecity/arcade.city dev",
     "format": "yarn workspaces run format",
     "patch-package": "patch-package",
-    "pods": "yarn workspace @arcadecity/city-mobile pods",
-    "postinstall": "expo-yarn-workspaces check-workspace-dependencies && rn-nodeify --install buffer,events,process,stream,util,inherits,fs,path,assert,crypto --hack --yarn && yarn workspace @arcadecity/city-mobile postinstall && yarn patch-package"
+    "pods": "yarn workspace @arcadecity/city-mobile pods"
   },
   "dependencies": {
     "moment": "2.29.1",
@@ -36,7 +35,8 @@
     "prettier": "2.3.1",
     "shx": "0.3.3",
     "typescript": "4.3.4",
-    "util-deprecate": "^1.0.2"
+    "util-deprecate": "^1.0.2",
+    "xmlhttprequest": "^1.8.0"
   },
   "workspaces": {
     "packages": [
